requirements_py38.txt
Flask==2.3.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.48
psycopg2-binary==2.9.5
colorlog
debugpy
flask-login
flask-assets
flask-cors
flask_login
pyyaml
addict
yapf
pytest